[cgo] refs #3 Refactor all functions by `make format`
