+++ lib/cgo/libfc_handle.go
