+++ lib/cgo/main.go
@@ -3,13 +3,9 @@ libskycoin shim
 */
 package main
 
-import (
-	
-)
-
 // #cgo CFLAGS: -I../../include
 import "C"
 
 func main() {
-	
+
 }
